Prevent flicker from analytics setup.
